refactor: unify cross-plugin helper copies under a drift guard
chat-lock, multipart, and the LRU cooldown helper existed as near-verbatim
copies across plugins (plugins ship as self-contained zips, so copying is
deliberate). The copies are now identical and pinned by
scripts/shared-copies.test.mjs, which fails the suite if they ever drift
apart — a fix in one copy must be applied to all.

The multipart copy in voice-transcription gains the CR/LF header
sanitization the other copies already had, and the cooldown helper moves
to a canonical cooldown.ts imported by after-hours, faq-bot, and
http-action.
