+++ typebot-connector/multipart.ts
@@ -11,6 +11,8 @@ export interface MultipartFile {
 
 // Assemble a multipart/form-data body as raw bytes so a binary attachment survives intact (a string body
 // would be UTF-8 re-encoded and corrupt non-UTF-8 bytes). Pure — no ctx.
+// NOTE: intentionally duplicated per plugin (plugins ship as self-contained zips) — keep all copies in
+// sync; scripts/shared-copies.test.mjs fails the build when they drift.
 export function buildMultipartBody(boundary: string, fields: MultipartField[], files: MultipartFile[]): Buffer {
   const parts: Buffer[] = [];
   for (const f of fields) {
